Refactor to helper module
