+++ lib/get-user-status.js
@@ -2,8 +2,8 @@
 const { URL } = require("url");
 const listify = require("listify");
 const config = require("../config.json");
-const workstreamsData = require("../sg/db.json").workstreams;
 const gitHubAPI = require("./helpers/github.js").api;
+const { workstreamFromRepo } = require("./helpers/workstreams.js");
 const jsonGitHubDatabase = require("./helpers/json-github-database.js");
 const html = require("escape-goat").htmlEscape;
 
@@ -21,7 +21,7 @@ module.exports = async (submitterGitHubID, repoName) => {
     if (individual.info.gitHubID.toLowerCase() === submitterGitHubID.toLowerCase()) {
       if (individual.verified) {
         if (individual.workstreams === "all" ||
-            individual.workstreams.includes(getWorkstream(repoName).id)) {
+            individual.workstreams.includes(workstreamFromRepo(repoName).id)) {
           return statusIndividual(submitterGitHubID, repoName);
         }
         return statusIndividualNoWorkstreams(submitterGitHubID, repoName);
@@ -49,7 +49,8 @@ module.exports = async (submitterGitHubID, repoName) => {
     }
 
     for (const entity of verifiedEntities) {
-      if (entity.workstreams === "all" || entity.workstreams.includes(getWorkstream(repoName).id)) {
+      if (entity.workstreams === "all" ||
+          entity.workstreams.includes(workstreamFromRepo(repoName).id)) {
         return statusEntity(submitterGitHubID, repoName, entity);
       }
     }
@@ -78,7 +79,7 @@ function statusIndividual(submitterGitHubID, repoName) {
 }
 
 function statusIndividualNoWorkstreams(submitterGitHubID, repoName) {
-  const workstreamName = getWorkstream(repoName).name;
+  const workstreamName = workstreamFromRepo(repoName).name;
   const repo = config.publicDataRepo.owner + "/" + config.publicDataRepo.name;
 
   return {
@@ -105,7 +106,7 @@ function statusIndividualUnverified(submitterGitHubID, repoName) {
 }
 
 function statusEntity(submitterGitHubID, repoName, entity) {
-  const workstreamName = getWorkstream(repoName).name;
+  const workstreamName = workstreamFromRepo(repoName).name;
 
   return {
     state: "success",
@@ -134,7 +135,7 @@ function statusEntitiesUnverified(submitterGitHubID, repoName, entities) {
 
 function statusEntityNoWorkstreams(submitterGitHubID, repoName, entities) {
   const { orgs, word, pronoun, pronoun2, doVerb } = entityVerbiage(entities);
-  const workstreamName = getWorkstream(repoName).name;
+  const workstreamName = workstreamFromRepo(repoName).name;
 
   return {
     state: "pending",
@@ -177,19 +178,4 @@ async function isOrgMember(org, username) {
   } catch (e) {
     return false;
   }
-}
-
-function getWorkstream(repoName) {
-  for (const workstream of workstreamsData) {
-    for (const standard of workstream.standards) {
-      if (getShortname(standard.href) === repoName) {
-        return workstream;
-      }
-    }
-  }
-  throw Error(`${repoName} not associated with a Workstream.`);
-}
-
-function getShortname(standardHref) {
-  return standardHref.slice("https://".length, standardHref.indexOf("."));
 }
