+++ lib/update-pr.js
@@ -3,10 +3,7 @@ const { BadRequest, Forbidden } = require("http-errors");
 const gitHubAPI = require("./helpers/github.js").api;
 const getUserStatus = require("./get-user-status.js");
 const config = require("../config.json");
-const workstreamsData = require("../sg/db.json").workstreams;
-
-const repos = workstreamsData.flatMap(workstream => workstream.standards)
-                             .map(standard => getShortname(standard.href)).join("|");
+const { repos } = require("./helpers/workstreams.js");
 
 module.exports = async url => {
   const prLocation = getPRLocation(url);
@@ -29,7 +26,7 @@ module.exports = async url => {
 function getPRLocation(url) {
   // URL will have been parsed and serialized by now, so this is reasonable to do.
   const re = new RegExp(
-    `^https://github.com/${config.specOrg}/(${repos})/pull/([1-9][0-9]*)(?:/[^/]+)?(?:#.*)?$`);
+    `^https://github.com/${config.specOrg}/(${repos.join("|")})/pull/([1-9][0-9]*)(?:/[^/]+)?(?:#.*)?$`);
   const match = re.exec(url);
 
   if (match === null) {
@@ -43,8 +40,4 @@ function getPRLocation(url) {
     repo: match[1],
     pull_number: Number(match[2])
   };
-}
-
-function getShortname(standardHref) {
-  return standardHref.slice("https://".length, standardHref.indexOf("."));
 }
